# BitWaves.JudgeBoard/docs/swagger/bitwaves-judgeboard-v1.yaml
swagger: '2.0'
info:
  description: BitWaves Judge Board API documentation
  version: 1.0.0
  title: BitWaves Judge Board API Documentation
  contact:
    email: msrlancern@126.com
...
paths:
...
  /judges:
    get:
      summary: Get all active judge nodes online
      description: Get all active judge nodes online
      tags:
        - judges
      responses:
        200:
          description: Operation completed successfully
          schema:
            type: array
            items:
              $ref: '#/definitions/JudgeNodeInfo'
    patch:
      summary: Patches the information about a judge node.
      description: Patches the information about a judge node. This is typically called a heartbeat. The requests must
          be made through a HTTPS session with the client holding valid judge node certificate.
      tags:
        - judges
      parameters:
        - in: body
          name: data
          description: Heartbeat information.
          required: true
          schema:
            $ref: '#/definitions/HeartbeatInfo'
      responses:
        200:
          description: Operation completed successfully.
        400:
          description: Invalid request data.
        401:
          description: No Authorization header are given or the authorization information has already expired.
        403:
          description: Invalid authorization information.
  /judges/{address}/block:
    put:
      summary: Change the block flag of a specified judge node.
      description: Change the block flag of a specified judge node.
      tags:
        - judges
      parameters:
        - in: path
          name: address
          required: true
          type: string
          description: The address of the judge node
        - in: query
          name: blocked
          required: false
          type: boolean
          description: Whether the judge node is blocked
          default: true
      responses:
        200:
          description: Operation completed successfully.
        400:
          description: Invalid request data.
        404:
          description: The judge node specified does not exist.
...
definitions:
...
  HeartbeatInfo:
    type: object
    required:
      - cpuUsage
      - cores
      - totalPhysicalMemory
      - freePhysicalMemory
      - swapFileSize
      - cachedSwapSpace
    properties:
      cpuUsage:
        type: number
        description: The overall CPU usage ratio of the judge node
        minimum: 0
        maximum: 1
        example: 0.8
      cores:
        type: number
        description: The number of CPU cores installed on the judge node
        minimum: 1
        example: 4
      totalPhysicalMemory:
        type: number
        description: Total amount of physical memory installed on the judge node, in megabytes.
        minimum: 1
        example: 2048
      freePhysicalMemory:
        type: number
        description: Total amount of physical memory that is free to use on the judge node, in megabytes.
        minimum: 0
        example: 128
      swapFileSize:
        type: number
        description: The size of the swap file, in megabytes.
        minimum: 0
        example: 32
      cachedSwapSpace:
        type: number
        description: The size of the swap space size, in megabytes.
        minimum: 0
        example: 21
  PerformanceInfo:
    type: object
    required:
      - cpuUsage
      - cores
      - totalPhysicalMemory
      - freePhysicalMemory
      - swapFileSize
      - cachedSwapSpace
    properties:
      cpuUsage:
        type: number
        description: The overall CPU usage ratio of the judge node
        minimum: 0
        maximum: 1
        example: 0.8
      cores:
        type: number
        description: The number of CPU cores installed on the judge node
        example: 4
      totalPhysicalMemory:
        type: number
        description: Total amount of physical memory installed on the judge node, in megabytes.
        example: 2048
      freePhysicalMemory:
        type: number
        description: Total amount of physical memory that is free to use on the judge node, in megabytes.
        example: 128
      swapFileSize:
        type: number
        description: The size of the swap file, in megabytes.
        example: 32
      cachedSwapSpace:
        type: number
        description: The size of the cached swap space, in megabytes.
        example: 21
  JudgeNodeInfo:
    type: object
    description: Provide information about an active judge node
    required:
      - address
      - lastHeartBeat
      - lastSeen
      - queuedSubmissions
      - isBlocked
      - performance
    properties:
      address:
        type: string
        description: The IP address of the judge node
        example: "127.0.0.1"
      lastHeartBeat:
        type: string
        description: Timestamp of the last heart beat packet sent by the judge node
        example: "2019-12-01T19:00:24Z"
      lastSeen:
        type: string
        description: Timestamp of the last request sent by the judge node
        example: "2019-12-01T19:00:24Z"
      queuedSubmissions:
        type: string
        description: Number of queued submissions currently dispatched to the judge node
        example: 2
      isBlocked:
        type: boolean
        description: Whether to prevent furthur dispatching submissions to the judge node.
        example: false
      performance:
        $ref: '#/definitions/PerformanceInfo'
        description: Performance related information of the judge node
...